requirements.txt
python-dotenv
flask
# openai==1.35.15
requests
streamlit
beautifulsoup4
streamlit-chat